+++ R/coerce.R
@@ -1,11 +1,13 @@
-sql_data <- function(value, row.names = FALSE, warn = FALSE) {
+sql_data <- function(value, conn, row.names = FALSE, warn = FALSE) {
   row.names <- compatRowNames(row.names)
   value <- sqlRownamesToColumn(value, row.names)
 
   value <- factor_to_string(value, warn = warn)
   value <- string_to_utf8(value)
-  value <- posixlt_to_posixct(value)
+  value <- difftime_to_hms(value)
+  value <- posixlt_to_posixct(value, conn@timezone)
   value <- numeric_to_finite(value)
+  value <- date_to_double(value)
   value
 }
 
@@ -30,9 +32,29 @@ string_to_utf8 <- function(value) {
   value
 }
 
-posixlt_to_posixct <- function(value) {
-  is_posixlt <- vlapply(value, inherits, "POSIXlt")
-  value[is_posixlt] <- lapply(value[is_posixlt], as.POSIXct)
+difftime_to_hms <- function(value) {
+  is_difftime <- vlapply(value, inherits, "difftime")
+  # https://github.com/tidyverse/hms/issues/84
+  value[is_difftime] <- lapply(value[is_difftime], function(x) {
+    mode(x) <- "double"
+    hms::as_hms(x)
+  })
+  value
+}
+
+posixlt_to_posixct <- function(value, timezone) {
+  is_posixlt <- vlapply(value, inherits, "POSIXt")
+  value[is_posixlt] <- lapply(value[is_posixlt], function(x) {
+    x <- as.POSIXct(x)
+
+    # The database expects times as local time
+    if (timezone != "UTC") {
+      x <- lubridate::with_tz(x, timezone)
+      x <- lubridate::force_tz(x, "UTC")
+    }
+
+    x
+  })
   value
 }
 
@@ -40,6 +62,15 @@ numeric_to_finite <- function(value) {
   is_numeric <- vlapply(value, is.numeric) & !vlapply(value, is.integer)
   value[is_numeric] <- lapply(value[is_numeric], function(x) {
     x[!is.finite(x)] <- NA
+    x
+  })
+  value
+}
+
+date_to_double <- function(value) {
+  is_date <- vlapply(value, inherits, "Date")
+  value[is_date] <- lapply(value[is_date], function(x) {
+    mode(x) <- "double"
     x
   })
   value