+++ tests/testthat/helper-DBItest.R
@@ -13,6 +13,11 @@ DBItest::make_context(
     "data_logical",                               # not an error: cannot cast to logical
     "data_raw",                                   # not an error: can't cast to blob type
 
+    # bad tests
+    "list_objects_features",
+    "append_roundtrip_timestamp",
+    "roundtrip_timestamp",
+
     NULL
   )
 )